test: finish the wasip2 split so every package's tests build
The move left internal/component/instance's test binary broken: five files
used WithWASI, which had left the package. Resolving it turned out to be
mostly deletion, because four of the five passed an EMPTY WASIConfig -- they
were not observing WASI at all, they were making a guest's imports resolve,
which the graph engine's trap-stub fallback already does. Dropping the
option is the whole fix for stackful_task, wast_conformance, and
wast_async_conformance.

The genuinely mixed ones were split along what each test is actually about:

  - TestCompileCache_HelloPrintsHelloWorld needs a WASI host to see what the
    cached component printed, so it moves to wasip2. Its cache-entry
    assertions do not survive the move (they read cache internals) and did
    not need to: TwoHelloLiveShareShims covers entry counting.
  - TestCompileCache_TwoHelloLiveShareShims reads cache internals, so it
    stays -- and drops its stdout capture, since counting cache entries
    never needed a WASI host to begin with.
  - resource_drop_async_test.go follows its fixture to wasip2.
  - The two multi-component tests that had accreted into real_hello_test.go
    are about the ENGINE running several components on one Runtime, not
    about WASI, so they move back as multicomponent_test.go.

Full suite green: 75 packages. The 35 differential conformance fixtures
still match their wasmtime goldens from imports/wasip2. Coverage 93.5% for
wasip2 and 87.8% for instance -- the latter is down from 90.4% because ~9,100
lines of WASI tests left the package while the engine's own line count
stayed, not because anything stopped being tested. make lint is 115, the
main baseline exactly.
