+++ internal/component/instance/wast_conformance_test.go
@@ -109,7 +109,7 @@ func runWastSuite(t *testing.T, suite string) {
 			if in != nil {
 				in.Close(ctx)
 			}
-			opts := WithWASI(WASIConfig{})
+			var opts []Option
 			if suite == "resources" {
 				// wasmtime/resources.wast imports a `host` instance that exports
 				// a resource type + its constructor/methods -- the embedder (the
